pyproject.toml
[tool.poetry]
name = "instagram"
version = "0.1.0"
description = ""
authors = ["jakhongir <ganiyevuzb@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
drf-yasg = "^1.21.5"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.2.2"
pillow = "^10.0.0"
psycopg2-binary = "^2.9.6"
django-jazzmin = "^2.6.0"
django-debug-toolbar = "^4.1.0"
python-dotenv = "^1.0.0"
daphne = "^4.0.0"
twisted = {extras = ["http2", "tls"], version = "^22.10.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
